Emit connectionDetails (CD2) and make msgSerial a pointer (ably-js compat)
Surfaced by the embedding PoC's M2 (Node) track while proving the
unmodified ably-js SDK. ably-go and the .NET IO.Ably SDK tolerate both
gaps; ably-js does not.

- CONNECTED now carries connectionDetails (Ably CD2) — connectionKey,
  clientId, maxIdleInterval. ably-js treats a CONNECTED without it as a
  protocol error and refuses the connection.
- ProtocolMessage.MsgSerial becomes *int64. As int64 with omitempty it
  silently dropped msgSerial=0 — the first publish on every connection —
  from the ACK, so ably-js could not correlate the ack to its pending
  publish and the publish promise hung forever. A pointer lets a
  meaningful zero serialise; helpers Int64()/MsgSerialValue() keep call
  sites tidy. Existing tests migrate to the pointer API with assertions
  unchanged.

Tracked as TASK-68 and TASK-69. Full suite: go test ./... green.

+++ internal/protocol/encode_test.go
@@ -10,7 +10,7 @@ func TestRoundTrip(t *testing.T) {
 	original := &ProtocolMessage{
 		Action:       ActionConnected,
 		ConnectionID: "abc123def456",
-		MsgSerial:    42,
+		MsgSerial:    Int64(42),
 		Timestamp:    1700000000000,
 	}
